fix x86 build

+++ tests/unit/test_fuzz.cpp
@@ -110,7 +110,7 @@ void TestFuzz(uint32_t aSeconds)
 #endif
 
 #ifdef OPENTHREAD_MULTIPLE_INSTANCE
-    uint64_t otInstanceBufferLength = 0;
+    size_t otInstanceBufferLength = 0;
     uint8_t *otInstanceBuffer = NULL;
 
     // Call to query the buffer size
