+++ tests/unit/test_message_queue.cpp
@@ -193,7 +193,7 @@ void TestMessageQueueOtApis(void)
     otMessage message;
 
 #ifdef OPENTHREAD_MULTIPLE_INSTANCE
-    uint64_t otInstanceBufferLength = 0;
+    size_t otInstanceBufferLength = 0;
     uint8_t *otInstanceBuffer = NULL;
 
     // Call to query the buffer size
